prj/M2_Encoder/requirements.txt
setuptools<81
torch==2.8.0
pytorch_lightning==1.9.2
transformers==4.17
Pillow
tqdm
ipdb==0.13.7
einops==0.3.0
pyarrow
sacred==0.8.2
pandas
timm==0.4.12
torchmetrics
torchvision
fairscale==0.4.1
protobuf==3.15.1
numpy==1.26.4
scipy
opencv-python==4.11.0.86
ftfy
sentencepiece
pyomo
mup
openspg-nn4k==0.0.2b3
modelscope
openvino
openvino-tokenizers

